+++ lib/nas/cp/file.js
@@ -43,9 +43,14 @@ async function getFileHash(filePath) {
   throw new Error('get file hash error, target is not a file, target path is: ' + filePath);
 }
 
-async function isEmptyDir(dirPath) {
-  const files = await fs.readdir(dirPath);
-  if (!files.length) { return true; }
+async function isEmptyDir(path) {
+  const lstat = await fs.lstat(path);
+  if (lstat.isDirectory()) {
+    const dirs = await fs.readdir(path);
+    if (_.isEmpty(dirs)) {
+      return true;
+    }
+  }
   return false;
 }
 
@@ -73,6 +78,7 @@ async function getFilePermission(filePath) {
 module.exports = {
   isDir,
   isFile,
+  isEmptyDir,
   getFileHash,
   zipWithArchiver,
   isEmptyDir,
