+++ lib/package/template.js
@@ -9,6 +9,7 @@ const zip = require('../package/zip');
 const definition = require('../definition');
 
 const { red, green } = require('colors');
+const { isEmptyDir } = require('../nas/cp/file');
 const { generateRandomZipPath } = require('../utils/path');
 
 const _ = require('lodash');
@@ -360,17 +361,6 @@ async function zipToOss(ossClient, srcPath, ignore, zipName = 'code.zip') {
 
   await fs.remove(randomDir);
   return objectName;
-}
-
-async function isEmptyDir(path) {
-  const lstat = await fs.lstat(path);
-  if (lstat.isDirectory()) {
-    const dirs = await fs.readdir(path);
-    if (_.isEmpty(dirs)) {
-      return true;
-    }
-  }
-  return false;
 }
 
 async function uploadLocalNasDir(ossClient, baseDir, serviceNasMapping) {
